Implementation of global variable 'globaluploadrate', helper method set_upload_timeout and reset_vessel_timeout, the commands 'set uploadrate speed' and 'show uploadrate', and modifications to 'upload' and 'run' commands in response to ticket #892.
In order to help resolve the issue of frequent vessel timeouts during file uploads, users can set the upload rate which seash will use to temporarily modify the timeout count during 'upload' and 'run' commands. Thus, when uploading a file, seash will temporarily set the timeout count to be whatever value (file size / globaluploadrate) results in, if larger than the current globalseashtimeout value. Upload rate is in bytes per second, and default is 10240 bytes per second (10 kb/sec). Users with slower connections can set a lower uploadrate to allow for more time for a file to upload before timing out. This does not, however, modify a user's connection speed in any manner, and the users are responsible for having a general idea of what kind of connection they have and the upload rate value that works for them.