seash.py has been restructured to clean up the command loop's code and allow for easier implementation of new commands. Much of the coding has been split up into separate files for easier accessibility and organization when a change needs to be made. Also, a command dictionary of dictionaries has been implemented as a data structure to keep track of the available commands for seash.
seash.py now only holds the command loop for taking in user input, as 
well as the temporary file name tab completer.

command_callbacks.py holds all of the code associated with each command 
that gets executed when the user calls for it.

seash_helper.py contains the helper functions and methods that operate 
on a target. Basically, all the declared methods that were coded before 
the actual command loop, and any other functions commonly used between 
certain commands.

seash_global_variables.py holds the original global variables in the 
original seash.py, and seash_exceptions.py holds all the custom coded 
exceptions that were added and used throughout seash.

seash_dictionary.py holds the newly implemented command dictionary, 
command parser, and command dispatcher. Details are explained in the 
comments, but basically the parser takes the user's input and checks if 
it follows a valid chain of sub-dictionaries within the command 
dictionary, while the dispatcher iterates through the passed chain of 
sub-dictionaries and executes the method associated with it.